refine sharing
