+++ ch01_bootsector/common.mk
@@ -106,7 +106,7 @@ endif
 
 # CI_PIPE_LINE_START
 FILES_SOURCE_COMMOM_PREFIX?=$(BASE_DIR)$(LANG_SUFFIX)/
-FILES_SOURCE_DEPENDENCIES?=$(FILES_SOURCE_COMMOM_PREFIX)common_$(TITLE).$(LANG_SUFFIX) $(FILES_SOURCE_COMMOM_PREFIX)common_bios.$(LANG_SUFFIX)
+FILES_SOURCE_DEPENDENCIES?=$(FILES_SOURCE_COMMOM_PREFIX)common_bios.$(LANG_SUFFIX)
 FILE_SOURCE?=$(BOOTSECTOR).$(LANG_SUFFIX)
 
 FILE_LANG_TO_C_FINAL_RESULT=$(BOOTSECTOR).c
