+++ ch01_bootsector/v/.vmodules/cache/README.md
@@ -1,0 +1,4 @@
+This folder contains cached build artifacts from the V build system.
+You can safely delete it, if it is getting too large.
+It will be recreated the next time you compile something with V.
+You can change its location with the VCACHE environment variable.
