+++ ch01_bootsector/v/common_bios.v
@@ -1,6 +1,6 @@
 module common_bios
 
-pub fn bios_clear_screen() {}
-pub fn bios_set_cursor_pos_p_row_col(row u8, col u8) {}
-pub fn bios_set_print_color_p_color() {}
-pub fn bios_print_string_p_msg(msg &char) {}
+pub fn asm_bios_clear_screen() {}
+pub fn asm_bios_set_cursor_pos(row u8, col u8) {}
+pub fn asm_bios_set_print_color(c u8) {}
+pub fn asm_bios_print_char(c u8) {}
