+++ ch01_bootsector/v/common_prefix.v.h
@@ -1,3 +1,7 @@
 #include "../c/common_prefix.h"
+
 #define _SLIT(x) x
 typedef char string[];
+typedef unsigned char u8;
+
+BEGIN_ASM_BOOTSECTOR;
