+++ ch02_timer/c/bootsector.c
@@ -38,4 +38,4 @@ BEGIN_TIMER_HANDLER;
 
 END_TIMER_HANDLER;
 
-END_BOOTSECTOR
+END_ASM_BOOTSECTOR
